API change of _io module in datasets subpackage to module io
