+++ docs/conf.py
@@ -28,7 +28,7 @@ sys.path.insert(0, os.path.abspath("sphinxext"))
 
 import watex 
 from watex.utils._packaging_version import parse
-
+from watex._docstring import wx_rst_epilog
 
 try:
     # Configure plotly to integrate its output into the HTML pages generated by
@@ -124,42 +124,7 @@ autodoc_typehints = "none"
 autodoc_typehints_format = "short"
 
 # Allow shorthand references for main function interface
-rst_prolog = """
-.. currentmodule:: watex
-"""
-
-# Define replacements (used in whatsnew bullets)
-
-rst_epilog = """
-
-.. role:: raw-html(raw)
-   :format: html
-   
-.. |ohmS| replace:: Pseudo-area of the fractured zone 
-.. |sfi| replace:: Pseudo-fracturing index 
-.. |VES| replace:: Vertical Electrical Sounding 
-.. |ERP| replace:: Electrical Resistivity Profiling 
-.. |MT| replace:: Magnetotelluric 
-.. |AMT| replace:: Audio-Magnetotellurics 
-.. |CSAMT| replace:: Controlled Source |AMT| 
-.. |NSAMT| replace:: Natural Source |AMT| 
-.. |EM| replace:: electromagnetic
-.. |EMAP| replace:: |EM| array profiling
-
-.. |Open Source? Yes!| image:: https://badgen.net/badge/Open%20Source%20%3F/Yes%21/blue?icon=github
-   :target: https://github.com/WEgeophysics/watex
-   
-.. |License BSD| image:: https://img.shields.io/github/license/WEgeophysics/watex?color=b&label=License&logo=github&logoColor=blue
-   :alt: GitHub
-   :target: https://github.com/WEgeophysics/watex/blob/master/LICENSE
-   
-.. |simpleicons git| image:: https://img.shields.io/badge/--F05032?logo=git&logoColor=ffffff
-   :target: http://git-scm.com 
-   
-.. |DOI| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.7744732.svg
-   :target: https://doi.org/10.5281/zenodo.7744732
-   
-""" 
+rst_prolog = wx_rst_epilog
 # noqa
 
 # -- Options for HTML output -------------------------------------------------
@@ -227,7 +192,7 @@ html_theme_options = {
         "json_url": json_url, 
         "version_match": switcher_version,
      },
-     "check_switcher": False,      
+     "check_switcher": True,      
     "show_prev_next": False,
     "navbar_start": ["navbar-logo", "version-switcher"],
     "navbar_end":  ["navbar-icon-links"], # [ "theme-switcher"]
