+++ docs/whatsnew/v0.1.0.rst
@@ -8,20 +8,20 @@ occured in :term:`Bagoue region`.
 
 Features
 ~~~~~~~~~~~~
-- Upload the first commit of the project that implemented the scripts to handle the case history problem. 
+- |Feature| Upload the first commit of the project that implemented the scripts to handle the case history problem. 
 
-- Developed :class:`watex.cases.BaseSteps` for processing the data at once.  Additionally, the default organization 
+- |Feature| Developed :class:`watex.cases.BaseSteps` for processing the data at once.  Additionally, the default organization 
   and scaling with different figure sizes has been improved.
 
-- Added the ``feature_transformation`` option to :mod:`watex.transformers` to automate the data transformation using 
+- |Feature| Added the ``feature_transformation`` option to :mod:`watex.transformers` to automate the data transformation using 
   :class:`watex.transformers.StratifiedWithCategoryAdder`.
 
-- Added the ability to seed the random number generator for the bootstrap used to define error bars in several plots especially 
+- |Feature| Added the ability to seed the random number generator for the bootstrap used to define error bars in several plots especially 
   in ``MetricPlot`` which was renamed to :class:`watex.view.EvalPlot` in the ``v0.1.3`` release and :class:`watex.view.QuickPlot`. 
 
-- Improved support for grouping observations based on pandas index information in categorical features plots. Should be auto-discarded 
+- |Feature| Improved support for grouping observations based on pandas index information in categorical features plots. Should be auto-discarded 
   using the powerfull function :func:`watex.utils.to_numeric_dtypes`. The later automates the data types conversion and outputs either 
   the numerical, categorical features or both.
 
-- Test the :term:`flow-rate` prediction using the support vector machines algorithmns from the defined DC-features. 
+- |Enhancement| Test the :term:`flow-rate` prediction using the support vector machines algorithmns from the defined DC-features.
